refactor: deprecate `encryptKey`, `decryptKey`, `sha256hash`
